Display owners details even when owners.expandGroups == false
When `owners.expandGroups = false` then neither groups nor accounts are
expaned to members (in case of groups) or account ids (in terms of
accounts) in the plugins REST API response. In addition, owner can be
configured as:
* email - in this case the `@domain` is dropped
* full user name - in this case full user name is conveyed

Considering the above when owner details are about to be displayed:
* check if it starts with `group/` prefix and display it as

  Group: {name} [copy_icon]

* search change's accounts through either domain dropped email address
  or full name and finally fallback to display

  {name} [copy_icon]

Note that either account name or group name is copied when button is
clicked.

Bug: Issue 379269836
Change-Id: I7c87dd9b91ab45045536f86b80b0ec7c99ba3613