Ensure that `Owned Files` is not switchable when no files are owned
When plugin registers itself for tab display even if `nothing` is
returned Gerrit still shows the pointer and allows clicking at it,
redirecting to the empty tab when done.

Traverse through the shadowRoots down to the tab and disable/enable
it when needed. As a result it is no longer possible to select an empty
tab without content.

Change-Id: Ib9da5fe0ed8ba564f8d62de2f8b13179ae232a4e