Don't display `Owned Files` when not current PS is viewed
The `owners` plugin's REST API returns files/owners status only for the
current_revision regardless of what is passed as the revision parameter.
Considering that it is better to not show `Owned Files` tab then display
different list of owned files as it may differ between PS'.

This behaviour is consistent with showing owners status on the `Files`
tab - it is not shown for the current_revision.

Bug: Issue 380110854
Change-Id: Id8fbcc1a6a6b8a52658e4f66a2fd69dd1960a917