Show `Owned Files` even when all files were approved
The `Owned Files` tab was introduced before `files_approved` property
was introduced to the owners plugin REST API response. As a result when
user approved change the `Owned Files` tab was no longer visible.

This change ensures that tab is visible even if all files are approved.
New functionality was covered in unit tests.

Bug: Issue 377723880
Change-Id: I2164287960a14fd833c0c0f27a53c8eca41d8901