Show owners files statuses and `Owned Files` tab in replacment mode
When `has:approval_owners` predicate is not configured for any project
in the hierarchy but `onwers.submitRequirements = true` is set in
`owners.config` instead then OWNERS file rules are still evaluted,
however, they are reported only in form of
`owners~OwnersSubmitRequirement` submit record.

Ensure that owners files statues and `Owned Files` tab are visible for
the corresponding changes. Note that common
`hasOwnersSubmitRequirementOrRecord` function was introduced (in
`owners-service.ts`) and it is being re-used consistently for both
`Files` and `Owned Files` tabs. It was also covered with unit tests.

Bug: Issue 381679562
Change-Id: I74c12d7cf88240736afdd38fab83fe22538a20cb