Fix misleading log message when adding reviewers
Commit I94576ca11 fixed an issue where an account already added as a
reviewer would be re-added to the change. However, this introduced a bug
where the warning message:

  "Not adding account X as reviewer to change Y because the associated
ref is not visible"

was incorrectly logged when an account was already a reviewer, even if
visibility was not the issue.

Ensure the warning is only logged when a reviewer is not added due to
visibility restrictions, and not when they are simply already assigned.

Bug: Issue 348598501
Change-Id: I7f15643634abc51a156861163ac32a7a6bfd4eb5