Default owner label config to max label value
Now that label.functions are defaulted to NoBlock and can no longer be
set to blocking values we should not rely on evaluating them to decide
if a change is approved or not. Default to using the maximum value for
the configured label if a value is not set.

Currently Code-Review label with no vote is used as default approval
score, default to Code-Review +2.

Bug: Issue 402138566
Change-Id: Id9b2198a330e711614f2e129971faa747970f852