Default Code-Review label approval score to +2
Currently the owners plugin defaults approval to Code-Review, 0 if
nothing is specified and the label.function is set to a non blocking
value.

It makes sense for the default Code-Review label approval score to be
set to +2 as that's the default in Gerrit core.

This can still be overwritten as before.

Bug: Issue 402138566
Change-Id: I5deb3ee6e14a861bc121dd6a67a492392491e749