Default Code-Review label approval score to +2
Currently the owners plugin defaults approval to Code-Review, 0 if
nothing is specified and the label.function is set to a non blocking
value.
This is explicitly in contrast with what's stated in the docs.

The documentation specifies that the default Code-Review label approval
score is set to +2, so the code should respect that.

This can still be overwritten as before.

Bug: Issue 402138566
Change-Id: I5deb3ee6e14a861bc121dd6a67a492392491e749