Share label representation between API and SR
Currently the owners~filesOwners api and the SubmitRequirement duplicate
logic to represent the definition of a label. These representations are
extremely similar but with subtle differences.

There is a third class, LabelDefinition, that should be used instead as
a unified representation of the approval label.

As there is no longer risk of the default label not having an assigned
score, the validation logic is simplified and unified across both
components. Before OwnersSubmitRequirement would use defaultLabelScore
if no vote was defined and the specified label function was NoBlock,
which is 0 while the API would use MaxPositive value of the same label,
leading to discrepancies between the API and the SR.

Bug: Issue 402138566
Change-Id: I2e5ebe9c28a7f60e4839b1dcc4db0aa4b0efcb02