Warn about off-line reindexing when using the owners' plugin
The owners plugin does not provide its predicates and functions
when running the off-line reindexing, however, the documentation did
not contain any note about it.

Add one specific note in the plugin's overview with the explicit
warning about the plugin NOT providing its functionality when
running off-line reindexing and the suggested command to use
for reindexing projects using the owners plugin.

Change-Id: I55c2028c72450f68604aee8d8d1e37d9d4ab851e