Drop support for label functions
Currently the SR takes into account the possibility of label.function
being defined, however the API doesn't. Label functions have been
deprecated since docs where created for them in, I25cd93c1, so we
shouldn't add support for them in the API. At the same time, the API and
the SR should be consistent.

The only solution is to drop support for label.functions in the SR part
of the owners plugin.

Change-Id: I04c87affe309be691e7a46cb869d0f197e20a059