Make label score compulsory in OWNERS file or owners.config
Currently it's possible to configure an approval label without a score
associated to it. This highlights some inconsistencies between the
has:approval_owners submit requirement logic and the API. The Submit
Requirement logic defaults to the default label score, usually 0, while
the API defaults to the max configured value for that score.

By forcing the definition of a label score and not having to account for
optional label scores in the code, we remove this inconsistency.

Bug: Issue 402138566
Change-Id: Id7c5c636c839f91469349203df909b761ff7c6ee