Make label score compulsory
Label score has been required for many version but we've tollerated it
being absent. This causes multiple problems as the API and the SR logic
differ when the label score is not defined.
The API logic looks for the max value configured for that label, while
the SR looks for the default value causing inconsistencies in the UI and
confusing end users.

Change-Id: Id7c5c636c839f91469349203df909b761ff7c6ee