Fix NPE in SyncReviewerManager
A NullPointerException occurred when attempting to auto-assign reviewers
synchronously, without explicitly registering the AutoassignModule in
gerrit.config. This prevented reviewers from being assigned to changes.

This was caused by a mismerge of 3.4 (Iff9b12), which needlessly kept
calling `addToAttentionSet` even when the full list of reviewers had
already been computed.

Bug: Issue 353080249
Change-Id: Idbeb9c5ce2235e4b97f2ca4daa18f0f8e234093f